feat(api): breaks v2 and pagination tests into separate file
