+++ api/registry/test/test_passport_get_score.py
@@ -174,78 +174,6 @@ class TestPassportGetScore:
 
             assert expected_pks == returned_pks
 
-    def test_get_scores_returns_first_page_scores_v2(
-        self,
-        scorer_api_key,
-        passport_holder_addresses,
-        scorer_community,
-        paginated_scores,
-    ):
-        limit = 2
-
-        client = Client()
-        response = client.get(
-            f"/registry/v2/score/{scorer_community.id}?limit={limit}",
-            HTTP_AUTHORIZATION="Token " + scorer_api_key,
-        )
-        response_data = response.json()
-
-        assert response.status_code == 200
-        assert response_data["prev"] == None
-
-        next_page = client.get(
-            response_data["next"],
-            HTTP_AUTHORIZATION="Token " + scorer_api_key,
-        )
-
-        assert next_page.status_code == 200
-
-        for i in range(limit):
-            assert (
-                response_data["items"][i]["address"]
-                == passport_holder_addresses[i]["address"].lower()
-            )
-
-    def test_get_scores_returns_second_page_scores_v2(
-        self,
-        scorer_api_key,
-        passport_holder_addresses,
-        scorer_community,
-        paginated_scores,
-    ):
-        limit = 2
-
-        client = Client()
-        page_one_response = client.get(
-            f"/registry/v2/score/{scorer_community.id}?limit={limit}",
-            HTTP_AUTHORIZATION="Token " + scorer_api_key,
-        )
-        page_one_data = page_one_response.json()
-
-        assert page_one_response.status_code == 200
-
-        page_two_response = client.get(
-            page_one_data["next"],
-            HTTP_AUTHORIZATION="Token " + scorer_api_key,
-        )
-        page_two_data = page_two_response.json()
-
-        assert page_two_response.status_code == 200
-
-        page_two_prev = client.get(
-            page_two_data["prev"],
-            HTTP_AUTHORIZATION="Token " + scorer_api_key,
-        )
-
-        assert page_two_prev.status_code == 200
-        assert page_two_prev.json() == page_one_data
-
-        for i in range(limit):
-            assert (
-                page_two_data["items"][i]["address"]
-                == passport_holder_addresses[i + limit]["address"].lower()
-            )
-
     def test_get_scores_request_throws_400_for_invalid_community(self, scorer_api_key):
         client = Client()
         response = client.get(
@@ -276,102 +204,6 @@ class TestPassportGetScore:
             response.json()["items"][0]["address"]
             == passport_holder_addresses[0]["address"].lower()
         )
-
-    def test_get_scores_filter_by_last_score_timestamp__gt(
-        self,
-        scorer_api_key,
-        passport_holder_addresses,
-        scorer_community,
-        paginated_scores,
-    ):
-        scores = list(Score.objects.all())
-        middle = len(scores) // 2
-        older_scores = scores[:middle]
-        newer_scores = scores[middle:]
-        now = datetime.datetime.now()
-        past_time_stamp = now - datetime.timedelta(days=1)
-        future_time_stamp = now + datetime.timedelta(days=1)
-
-        # Make sure we have sufficient data in both queries
-        assert len(newer_scores) >= 2
-        assert len(older_scores) >= 2
-
-        for score in older_scores:
-            score.last_score_timestamp = past_time_stamp
-            score.save()
-
-        # The first score will have a last_score_timestamp equal to the value we filter by
-        for idx, score in enumerate(newer_scores):
-            if idx == 0:
-                score.last_score_timestamp = now
-            else:
-                score.last_score_timestamp = future_time_stamp
-            score.save()
-
-        # Check the query when the filtered timestamo equals a score last_score_timestamp
-        client = Client()
-        response = client.get(
-            f"/registry/score/{scorer_community.id}?last_score_timestamp__gt={now.isoformat()}",
-            HTTP_AUTHORIZATION="Token " + scorer_api_key,
-        )
-        assert response.status_code == 200
-        assert len(response.json()["items"]) == len(newer_scores) - 1
-
-        # Check the query when the filtered timestamo does not equal a score last_score_timestamp
-        response = client.get(
-            f"/registry/score/{scorer_community.id}?last_score_timestamp__gt={(now - datetime.timedelta(milliseconds=1)).isoformat()}",
-            HTTP_AUTHORIZATION="Token " + scorer_api_key,
-        )
-        assert response.status_code == 200
-        assert len(response.json()["items"]) == len(newer_scores)
-
-    def test_get_scores_filter_by_last_score_timestamp__gte(
-        self,
-        scorer_api_key,
-        passport_holder_addresses,
-        scorer_community,
-        paginated_scores,
-    ):
-        scores = list(Score.objects.all())
-        middle = len(scores) // 2
-        older_scores = scores[:middle]
-        newer_scores = scores[middle:]
-        now = datetime.datetime.now()
-        past_time_stamp = now - datetime.timedelta(days=1)
-        future_time_stamp = now + datetime.timedelta(days=1)
-
-        # Make sure we have sufficient data in both queries
-        assert len(newer_scores) >= 2
-        assert len(older_scores) >= 2
-
-        for score in older_scores:
-            score.last_score_timestamp = past_time_stamp
-            score.save()
-
-        # The first score will have a last_score_timestamp equal to the value we filter by
-        for idx, score in enumerate(newer_scores):
-            if idx == 0:
-                score.last_score_timestamp = now
-            else:
-                score.last_score_timestamp = future_time_stamp
-            score.save()
-
-        # Check the query when the filtered timestamo equals a score last_score_timestamp
-        client = Client()
-        response = client.get(
-            f"/registry/score/{scorer_community.id}?last_score_timestamp__gte={now.isoformat()}",
-            HTTP_AUTHORIZATION="Token " + scorer_api_key,
-        )
-        assert response.status_code == 200
-        assert len(response.json()["items"]) == len(newer_scores)
-
-        # Check the query when the filtered timestamo does not equal a score last_score_timestamp
-        response = client.get(
-            f"/registry/score/{scorer_community.id}?last_score_timestamp__gte={(now - datetime.timedelta(milliseconds=1)).isoformat()}",
-            HTTP_AUTHORIZATION="Token " + scorer_api_key,
-        )
-        assert response.status_code == 200
-        assert len(response.json()["items"]) == len(newer_scores)
 
     def test_get_single_score_for_address_in_path(
         self,
